Removed MSVC warnings. These are warnings which show up when using Ceres. I deemed them more serious than the many which show up when building Ceres.

+++ include/ceres/local_parameterization.h
@@ -160,7 +160,7 @@ class CERES_EXPORT SubsetParameterization : public LocalParameterization {
                     double* x_plus_delta) const;
   virtual bool ComputeJacobian(const double* x,
                                double* jacobian) const;
-  virtual int GlobalSize() const { return constancy_mask_.size(); }
+  virtual int GlobalSize() const { return static_cast<int>(constancy_mask_.size()); }
   virtual int LocalSize() const { return local_size_; }
 
  private:
